tools/games/requirements.txt
pillow
requests
matplotlib
numpy
pynput